+++ engine/src/main/java/org/wickedsource/docxstamper/util/DocumentUtil.java
@@ -11,6 +11,7 @@ import org.docx4j.openpackaging.parts.relationships.RelationshipsPart;
 import org.docx4j.wml.*;
 import org.jvnet.jaxb2_commons.ppp.Child;
 import pro.verron.officestamper.api.OfficeStamperException;
+import pro.verron.officestamper.core.RunUtil;
 
 import java.util.*;
 import java.util.stream.Stream;
