+++ engine/src/main/java/org/wickedsource/docxstamper/util/ParagraphUtil.java
@@ -5,6 +5,7 @@ import org.docx4j.wml.ObjectFactory;
 import org.docx4j.wml.P;
 import org.docx4j.wml.R;
 import pro.verron.officestamper.api.OfficeStamperException;
+import pro.verron.officestamper.core.RunUtil;
 
 /**
  * Utility class for creating paragraphs.
